+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## ** V0.2.0  (18/01/2024) **
+- feat:
+  - agent in MQL5
+  - files utils
+  - new util method
+- test
+  - files utils
+  - forex client
+  - new config parameters
+
 ## ** V0.1.0  (05/01/2024) **
 - feat:
   - forex handle